+++ src/forsitan.cpp
@@ -37,7 +37,7 @@ void init(Plugin* p) {
 	p->addModel(modelRaucus);
 	p->addModel(modelTundo);
 	p->addModel(modelCartilago);
-	p->addModel(modelTurba);
+	p->addModel(modelScrupea);
 	// Any other plugin initialization may go here.
 	// As an alternative, consider lazy-loading assets and lookup tables when your module is created to reduce startup times of Rack.
 }